feat: support updating/replacing rulesets
This commit adds support for adding/modifying rulesets associated with a
folder.

Currently missing is removal of rules if the entire ruleset is removed
from the configuration file. A workaround for this is to not remove the
entire ruleset from the YAML configuration, but leave an explicit empty
list. We can/should fix this in the future.

WO-1290