fix(checkmk-client): allow more redirects on change activation
Checkmk uses redirects on the `wait-for-completion` endpoint for
activating pending changes, in their words: "to prevent timeouts".

To make sure we don't fail early, this commit modifies the
redirect-policy of reqwest to follow redirects for the activation
endpoint up to 100 times, whereas for other paths it will apply the
default redirect policy (which at the time of this change is 10
redirects).

WO-1291